refactor: rename filename userStore -> user-store
