+++ src/hooks/use-is-logged-in.ts
@@ -1,4 +1,4 @@
-import { userStore } from '@/store/userStore'
+import { userStore } from '@/store/user-store'
 
 const useIsLoggedIn = () => {
   const user = userStore((s) => s.user)
