+++ src/hooks/use-user-point-history.ts
@@ -1,5 +1,5 @@
 import { usersApi, usersQueryKeys } from '@/api/users'
-import { userStore } from '@/store/userStore'
+import { userStore } from '@/store/user-store'
 import { PointFilterType } from '@/types/points'
 import { useQuery } from '@tanstack/react-query'
