+++ src/hooks/use-user-status.ts
@@ -1,5 +1,5 @@
 import { usersApi, usersQueryKeys } from '@/api/users'
-import { userStore } from '@/store/userStore'
+import { userStore } from '@/store/user-store'
 import { useQuery } from '@tanstack/react-query'
 
 export const useUserStatus = () => {
