+++ src/store/api-server-mocking-store.ts
@@ -1,7 +1,7 @@
 import { Challenge, Team } from '@/api/challenges'
 import { create } from 'zustand'
 import { devtools, persist } from 'zustand/middleware'
-import { User } from './userStore'
+import { User } from './user-store'
 import { v4 } from 'uuid'
 import { ClientAddress } from '@/types/addresses'
 import { serverToClientAddress } from '@/lib/utils'
